+++ genomepy/providers/__init__.py
@@ -106,13 +106,15 @@ def online_providers(provider: str = None):
             logger.warning(str(e))
 
 
-def search(term: str or int, provider: str = None, size=False):
+def search(term: str or int, provider: str = None, exact=False, size=False):
     """
-    Search for a genome.
+    Search for term in genome names and descriptions (if term contains text. Case-insensitive),
+    assembly accession IDs (if term starts with GCA_ or GCF_),
+    or taxonomy IDs (if term is a number).
 
-    If provider is specified, search only that specific provider, else
-    search all providers. Both the name and description are used for the
-    search. Search term is case-insensitive and can contain regex.
+    If provider is specified, search only that specific provider, else search all providers.
+
+    Note: exact accession ID search on UCSC may return different patch levels.
 
     Parameters
     ----------
@@ -120,6 +122,8 @@ def search(term: str or int, provider: str = None, size=False):
         Search term, case-insensitive, allows regex.
     provider : str , optional
         Only search the specified provider (faster).
+    exact : bool, optional
+        term must be an exact match
     size : bool, optional
         Show absolute genome size.
 
@@ -129,7 +133,7 @@ def search(term: str or int, provider: str = None, size=False):
         genome name, provider and metadata
     """
     for p in online_providers(provider):
-        for row in p.search(term, size):
+        for row in p.search(term, exact, size):
             ret = list(row[:1]) + [p.name] + list(row[1:])
             yield ret