+++ tests/test_01_basic.py
@@ -84,14 +84,13 @@ def test_exceptions():
 
 def test_config():
     config = norns.config("genomepy", default="config/default.yaml")
-    assert len(config.keys()) == 5
+    assert len(config.keys()) == 6
 
 
-def test_manage_config(capsys):
+def test_manage_config(caplog, capsys):
     # make a new config
     genomepy.config.manage_config("generate")
-    captured = capsys.readouterr().out.strip()
-    assert captured.startswith("Created config file")
+    assert caplog.text.count("Created config file") == 1
 
     # check where it is found
     fname = os.path.expanduser("~/Library/Application Support/genomepy/genomepy.yaml")
@@ -112,8 +111,7 @@ def test_manage_config(capsys):
 
     # make a new config
     genomepy.config.manage_config("generate")
-    captured = capsys.readouterr().out.strip()
-    assert captured.startswith("Created config file")
+    assert caplog.text.count("Created config file") == 2
 
     # check if the mess was fixed
     genomepy.config.manage_config("show")